+++ internal/network/http_client.go
@@ -65,7 +65,6 @@ func InitHTTPClients() {
 			DialContext: (&net.Dialer{
 				Timeout:   30 * time.Second,
 				KeepAlive: 30 * time.Second,
-				Control:   setSocketBuffer,
 			}).DialContext,
 			MaxIdleConns:          1000,
 			MaxIdleConnsPerHost:   1000,